`loadFileData()` checks every csv their result type using columns or summarisedResult class as reference

`loadFileData()` checks every csv their result type using columns or summarisedResult class as reference

Usage

loadFileData(
  data,
  fileName,
  configData,
  resultsData,
  resultsColumns,
  databaseName
)

Arguments

data

named list with data.

fileName

the file to load

configData

config loaded from file

resultsData

the loaded file data

resultsColumns

the loaded file columns

databaseName

db name

Value

the list with data