`table1NumPar()` creates table 1 with number of participants from each database from attrition data

`table1NumPar()` creates table 1 with number of participants from each database from attrition data

Usage

table1NumPar(prevalence_attrition, incidence_attrition)

Arguments

prevalence_attrition

A dataframe with prevalence attrition data.

incidence_attrition

A dataframe with incidence attrition data.

Examples

table1NumPar(prevalence_attrition = ReportGenerator:::prevalence_attrition_latest,
             incidence_attrition = ReportGenerator:::incidence_attrition_latest)
#>                                                  CHUBX                
#>             analysis_step   reason             current_n    excluded  
#>             initial         Starting            2.15e+06              
#>                             population                                
#>             initial         Missing year of     2.15e+06    0         
#>                             birth                                     
#>             initial         Missing sex         2.15e+06    0         
#>             initial         Cannot satisfy      1.99e+06    1.62e+05  
#>                             age criteria                              
#>                             during the study                          
#>                             period based on                           
#>                             year of birth                             
#>             initial         No observation      7.14e+05    1.28e+06  
#>                             time available                            
#>                             during study                              
#>                             period                                    
#>             initial         Doesn't satisfy     7.14e+05    0         
#>                             age criteria                              
#>                             during the study                          
#>                             period                                    
#>             initial         Prior history       4.12e+05    3.02e+05  
#>                             requirement not                           
#>                             fulfilled during                          
#>                             study period                              
#>             initial         Not Female          2.06e+05    2.06e+05  
#>             initial         Not Male            2.06e+05    2.06e+05  
#>             prevalence      Starting            1.36e+05              
#>                             analysis                                  
#>                             population                                
#>             prevalence      Excluded due to     1.3e+05     5.16e+03  
#>                             prior event (do                           
#>                             not pass outcome                          
#>                             washout during                            
#>                             study period)                             
#>             prevalence      Not observed        1.3e+05    39         
#>                             during the                                
#>                             complete                                  
#>                             database                                  
#>                             interval                                  
#>                             No observation      1.36e+05    1.39e+05  
#>                             time available                            
#>                             after applying                            
#>                             age, prior                                
#>                             history and, if                           
#>                             applicable,                               
#>                             strata criteria                           
#>             incidence       Starting            1.36e+05              
#>                             analysis                                  
#>                             population                                
#>             incidence       Excluded due to     1.3e+05     5.16e+03  
#>                             prior event (do                           
#>                             not pass outcome                          
#>                             washout during                            
#>                             study period)                             
#>             incidence       Not observed        1.3e+05    39         
#>                             during the                                
#>                             complete                                  
#>                             database                                  
#>                             interval                                  
#> 
#> Column names: analysis_step, reason, current_n, excluded